refactor: move cross_correlate_particle_stack to utils
