+++ src/tt2dtm/backend/__init__.py
@@ -1,7 +1,8 @@
 """Submodule for computationally intensive backend functions."""
 
 from .core_match_template import core_match_template
-from .core_refine_template import core_refine_template, cross_correlate_particle_stack
+from .core_refine_template import core_refine_template
+from .utils import cross_correlate_particle_stack
 
 __all__ = [
     "core_match_template",
